+++ src/chunk_append/exec.c
@@ -449,7 +449,15 @@ choose_next_subplan_non_parallel(ChunkAppendState *state)
 static void
 choose_next_subplan_for_leader(ChunkAppendState *state)
 {
-	state->current = NO_MATCHING_SUBPLANS;
+	/*
+	 * If no workers got launched for this parallel plan
+	 * we have to let leader participate in subplan
+	 * execution.
+	 */
+	if (state->pcxt->nworkers_launched == 0)
+		choose_next_subplan_for_worker(state);
+	else
+		state->current = NO_MATCHING_SUBPLANS;
 }
 
 static void
@@ -616,6 +624,7 @@ chunk_append_initialize_dsm(CustomScanState *node, ParallelContext *pcxt, void *
 
 	state->choose_next_subplan = choose_next_subplan_for_leader;
 	state->current = INVALID_SUBPLAN_INDEX;
+	state->pcxt = pcxt;
 	state->pstate = pstate;
 }